ci,build: add generic mechanism for passing some Travis-CI envs
This mechanism has been roughly adapted from the way it's done for the
Linux build.
A file is used on the mount-point that is passed to the docker, and the
'inside_docker.sh' script will load it if found.

Also added 3 common/variables. Two of them are used in the Linux build.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>