Doxygen: put results in a version'ed subdirectory
this will put Doxygen into a version'ed subdirectory, so when we copy to
gh-pages, we will not write overtop of older versions (which people may
still need).

This also put things in a ./libiio dir, since we will soon be putting
python bindings side by side. (in the same version directory)

Signed-off-by: Robin Getz <robin.getz@analog.com>