remove strcpy, and replace with an iio_strlcpy
We used strcpy a few places, which opens up classic buffer overflow
issues so check those, and fix them.

https://cwe.mitre.org/data/definitions/120.html

This adds some internal error checking, as we are building up xml
representations of things, we now keep track of the remaining space
in the buffer, and don't go over if we have no space. (which in theory
should never happen).

Tested on Pluto and M2k to see if this introduces issues, and couldn't
find anything.

Signed-off-by: Robin Getz <robin.getz@analog.com>