flawfinder: mark some ignores for CWE-134 in debug.h
According to https://cwe.mitre.org/data/definitions/134.html describes
functions that accepts a format string as an argument, but the format
string originates from an external source as dangerous.

This gets tagged on all the IIO_DEBUG, IIO_INFO, IIO_WARNING, and
IIO_ERRRO functions that are in debug.h However, they are always called
internally from the library, have fixed format strings and can not be
modified externally, so mark them as safe.

Signed-off-by: Robin Getz <robin.getz@analog.com>