Cmake/Coverity: Force coverity to check _everything_
The current Cmake logic is set so that if things (like avahi) are not
installed, it doesn't build or link against it. On normal systems this
make sense.

I recently noticed that coverity builds (which happen on the coverity
servers, not on the Travis-CI servers), don't have external libraries
installed, and then a few options don't get configured/added, and
therefore, don't get checked.

Add an option to the Cmake which blindly adds things (which is fine,
since coverity doesn't actually compile/link things); so all backends
and options get checked.

Signed-off-by: Robin Getz <robin.getz@analog.com>