cmake: support old Cmake for -std=c99
on old verisons of Cmake, they don't support the way we are setting the
"C_STANDARD 99" (in set_target_properties), so check for old verison and
then set the CMAKE_C_FLAGS by hand.

This eliminates all the warnings that happen on the Zynq images we use
(2019R1) and some of the older CI images we support, and make it alot
easier to see where the real errors/warnings are.

Signed-off-by: Robin Getz <robin.getz@analog.com>