csharp: Control Structures should use curly braces {}
According to Microsoft, and Sonar C#, all Control structures (do, for,
foreach, while, if) should use curly braces. While not technically incorrect,
the omission of curly braces can be misleading, and may lead to the
introduction of errors during maintenance.

See CERT, EXP19-C Use braces for the body of an if, for, or while statement
https://www.securecoding.cert.org/confluence/x/1QGMAg

See CERT, EXP52-J Use braces for the body of an if, for, or while statement
https://www.securecoding.cert.org/confluence/x/3wHEAw

Signed-off-by: Robin Getz <robin.getz@analog.com>