fix issues pointed out by no-sign-compare removal
By default comparing signed with unsigned, the compiler converts the
signed value to unsigned, this can cause problems sometimes, as that
would mean -1 > 2U.

In channel.c, context.c, device.c the lengths of strings to a negative
number, so cast to a signed number in all cases.

In dns_sd.c, local.c, tests/iio_stresstest.c we had an iterator, that
was signed, that should have been unsigned (now it is).

In local.c:get_rel_timeout_ms we move the comparision to signed.

in local.c, while pointers are unsigned integers, subtraction of two
pointers are ptrdiff_t, or a signed integer. Since we are just measuring
length of a string, cast to unsigned.

Signed-off-by: Robin Getz <robin.getz@analog.com>