Move scan_block functions into iio.h
scan_block functions we originally required by MATLAB but these are
generally handy in other applications. Moving into the main header.

This eliminates the iio-wrapper.h file, as it is not needed anymore.
Users who want to use this in MATLAB will need to add:
before they include iio.h

Signed-off-by: Travis F. Collins <travis.collins@analog.com>
Signed-off-by: Robin Getz <robin.getz@analog.com>