tests: add a common function to parse and create iio_contexts
Some (older) implementations of getopt_long) do not protect against internal
buffer overflows, so copy argv to a temp str array, and parse most in a
common function. (-u -h -x -n, and -S).
  -h, --help Show this help and quit.
  -x, --xml  Use the XML backend with the provided XML file.
  -u, --uri  Use the context at the provided URI.
  -S, --scan Scan for available backends.
  -a, --auto Scan for available contexts and if only one is available use it.

The "default" for scan is "S", since "s" is used for many apps as sample
size. This means that iio_info now takes -s and -S as scan.

This also fixes both CWE-120 and CWE-20
https://cwe.mitre.org/data/definitions/120.html
https://cwe.mitre.org/data/definitions/20.html
and almost is a net decrease (348 insertions, 345 deletions over 9
files).

Signed-off-by: Robin Getz <robin.getz@analog.com>