Cmake: move where -Werrors are turned on in the Cmake
when we turned on -Werrors too soon, it would cause issues like this:

./build/CMakeFiles/CMakeTmp/CheckSymbolExists.c: In function ‘main’:
./build/CMakeFiles/CMakeTmp/CheckSymbolExists.c:8:11: error: ISO C forbids conversion of function pointer to object pointer type [-Werror=pedantic]
   return ((int*)(&libusb_get_version))[argc];
           ^
cc1: all warnings being treated as errors

The problem being that this isn't an error, but -Wpedantic and -Werror
cause it to be one. So, we shuffle down the place in the Cmake where we
set the -Werror, and shuffle up where we look for check_symbol_exists().

While we are here, we also cat the CMake error log on the CI to help
debug issues like this in the future.

Signed-off-by: Robin Getz <robin.getz@analog.com>