iio_attr: add error checking
When helping debug someone's shell scripts what used iio_attr on
Friday, we noticed that iio_attr just plain eats too many things,
which makes it hard to debug things when using -q (quiet).

So, make sure to print things when errors occur. Errors are now defined
as attempting to read/write a device/channel/buffer/attribute that
doesn't exist. or when a write error occurs.

Also exit with an exit error code, so that if you set -e in a shell
script, it quits right away.

Signed-off-by: Robin Getz <robin.getz@analog.com>