usb: if blank usb uri (eg 'usb:'), search for one, and use that
Similar to the IP context, where we accept 'ip:' and it does a dns-sd
search for a device, accept USB URIs with the format 'usb:' which will
create a scan context and use if if there is only one usb context
available. If there is more than one, reture an error condition.

Tested with zero, one and two usb devices, with various networking
contexts in ths system.

Signed-off-by: Robin Getz <robin.getz@analog.com>