coverity: handle strto[*] functions under|over-flows robustly
a recent change that we (I) made to things wasn't handling
under|over-flows properly. The man page for strtol states:

Since strtol() can legitimately return 0, LONG_MAX, or LONG_MIN on both
success and failure, the calling program should set errno to 0 before the
call, and then determine if an error occurred by checking whether errno
has a nonzero value after the call.

Since all strto* functions (strtoll, strtol, strtoul, strtod, strtof) are
similar - set errno to 0, and see if it is errno is set to ERANGE
when it's finished.

checked with 'grep -A 1 -B 1 strto * -R' in the repo
Review the entire code base, and do the same thing everywhere.

tested on x86 and Zynq + FMCOMMS3

Signed-off-by: Robin Getz <robin.getz@analog.com>