xml: Add capability to properly encode xml entities in context attributes
Per #519, xml entities (<, >, ", ', and &) were not encoded properly, and if
used in the libiio.ini file (which was added in 77568c71ec0d7b0162f0b488f261d6d9874f2e88 )
cause the libiio library to fail.

This fixed that by adding a generic way to encode things in xml.c and uses them in context.c

Right now, we only encode context attributes. In theory this could also effect:
 - attribute name
 - buffer-attribute name
 - channel id
 - context name
 - debug-attribute name
 - device id
 - scan-element index
Since most of those are encoded in file names, we shouldn't have those chars
anyway. Now that we know the issue, and understand the fix, if anyone sees
any issues, please feel free to report.

Tested on Zed + FMCOMMS3, with both 0.19 (with fix) and old libraries without fix.
As long as the the version which is reading the file has the fix - it works with
older libraries (is backwards compatible).

Signed-off-by: Robin Getz <robin.getz@analog.com>