iio_utils: rework/fix command line argument parsing
the rework and moving to common functions for argument parsing broke a
few things, since I thought that you could pass a incomplete optstring
and incomplete longopts to getopt_long(). However, this is not the case.
It broke long flags (--uri vs -u) and broke the position parsing that
happened deep inside getopt_long() somewhere.

In fact - this caused builds on MSVC to crash sometimes. The other thing
that caused MSVC to crash was if an flag that had an optional argument
was the last on on the command line (for example 'iio_info -a') we would
use a index that exceeded the actual memory allocations, so check for
that and don't do it.

Now, make sure to pass complete arguments to things, and fix the bounds
checking which seems to resolve all the issues on Linux and Windows.

Signed-off-by: Robin Getz <robin.getz@analog.com>