iiod: Make startup (with Avahi) more robust.
In certain embedded cases, iiod may start before networking or dbus is
fully alive on a system (like on Pluto and M2k).

This change ensures that when we are trying to connect to avahi, and
register on the network, we give things a few moments to stablize (re-trying)
before we fail.

tested on PlutoSDR, and Zedboard + FMCOMMS3

Signed-off-by: Robin Getz <robin.getz@analog.com>