travis-ci: avoid errors when uploading things to akami
I don't know who maintains this infrastructure - but they don't use the
same keys on all their mirrors, so we need to use -o "StrictHostKeyChecking no"
which defeats the purpose of secure anything.

But this will avoid errors when pushing things up.

Since this is only pushing things up, not pulling things down - this is
not risky.

Signed-off-by: Robin Getz <robin.getz@analog.com>