network: Add consume to client ops
The goal of this new callback is to consume all the __possible__ stalled
data in the buffers before doing a new transaction. The problem can occur
if a remote device is occupied doing something when a new transaction is
started (eg: read an attribute). In this case, most likely, our read will
timeout and, due to the nature of stream sockets, the next attribute we
try to read will get the stalled value from the previous, failed,
transaction.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>