iio: split iio_backend_ops into iio-backend.h
The iio_backend_ops will be split away from libiio core. So, move the
iio_backend_ops into another header.
This header will be available in the IIO backends, and the rest of the
libiio core internals will be hidden from IIO backends.

Right now, this split isn't doing too much [mostly moving some types
around]. And it's not yet ready to be used on it's own.
It's only being used in 'iio-private.h', in the same manner as before.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>