local: use API accessors for accessing devices from an IIO context
The aim is to make all libiio backends plugins, so they need to decouple a
bit from accessing internal IIO context objects, and channels.

For all the backends in libiio, this change replaces the access of IIO
devices count and objects with IIO API accessors to get the number of
devices and a reference for each device when iterating.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>