xml: use iio_context_destroy() on error path of iio_create_xml_context_helper()
The error path of the iio_create_xml_context_helper() has all the traits of
a iio_context_destroy() function. iio_context_destroy() does plenty of
null-checks when destroying data on a context object.

This change, uses the iio_context_destroy() function on the error path of
iio_create_xml_context_helper().

The intent is to hide the 'iio_context' type inside the context.c file for
better modularity.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>