iio: create iio_context_create_from_backend() helper and use it for local & xml
The creation of a context can be tied to an iio_backend object that acts as
a template for some context parameters.
Using an iio_backend object for IIO context creation allows us to hide the
assignments of the name, iio_backend_ops, description, private data and
possibly other parameters in the future.

For now, this ties it to the local & xml backends. The other backends seem
to be iiod client-type backends. Using the iio_backend object there
requires a bit more rework.

At a later point in time, the iio_context_create_from_backend() helper
should also do more validation for the backend object.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>