xml: split iio_populate_xml_context_helper() from create helper
The iio_create_xml_context_helper() function is used in the manner that it
allocates an IIO context, then the network,serial & usb backends override
the name, description & pdata.

This design is not ideal:
1. The description is overridden by each backend; this requires that each
backend free the previous description string, in order to not leak memory.
The logic in network.c seems a bit too complicated (for handling the
ctx->description, and serial.c forgets to free the description allocated by
this helper).
2. Similar, each backend needs to allocate it's own pdata. We can abstract
this away so that the 'sizeof_context_pdata' is used so that a memory block
is allocated and returned back to the backend. The alloc & free of the
memory block can be handled by iio_context_destroy()

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>