iio: backend: add get_description() backend op
This doesn't get rid of the 'ctx->description' object; that will still be
around and cache the description.

The get_description() splits the creation logic of the description into a
backend op. That way, we can think of splitting the creation of an IIO
context into an 'alloc()' and 'init()' part where:
* an IIO context is allocated first in libiio core
* it's backend init() routine is called
* it's backend get_description() routine is called

The idea of the 'get_description()' op is to return a pointer, that libiio
core will manage/free. Hence the return is 'char *' vs 'const char *'.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>