ci,travis: remove lsb package requirement
To get the lsb_release command working some packages need to be
pre-installed.
The docker images come pretty bare, so they need to be installed before.

This can complicate things, because we need to make sure that all builds
down the pipeline install this package.

The utilities that wrap lsb_release can be re-implemented with simple shell
commands and parsing of /etc/os-release.
This change does that.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>