.github/actions: add initial build.yml file
Currently just does a build, an only on PRs to save time/resources.
Deployments will be re-added in the next commit series.

On push events only 'staging/**' branches are built.
This is an opt-in mechanism, since we may not want to build everything,
which is a waste of resources.

Dropped CentOS 6 builds during migration; we can re-add it; but it's old
enough.
The only MacOS X images on GH actions are version 10.5 and 11.0

Suffixed builds with CPU architecture `_x86_64`. We may want to add `_arm`
or `_arm64` later.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>