iiod/init: add a script which creates /etc/libiio.ini and bolt it in
This add a small script /usr/sbin/iiod_create_ini which can create a
/etc/libiio.ini file. This is connected to the startup scripts, and run
before iiod, so that things are populated on time.

Add a small script that will configure USB gadget (if wanted) so that
IIO devices are available over USB, and that iiod starts with the proper
options to do this.

Re-work a little of the CMake so that this is included during an
install.

Update the doc (the README_BUILD.md) so these options are clear for
people.

This is the first stage of #593

Signed-off-by: Robin Getz <robin.getz@analog.com>