ci,azure-pipelines: add coverity job
This change adds job for the Coverity scan project. There doesn't seem to
be an environment variable (at first glance) for the repo slug.
So, the name of the repo was added directly in the yaml file.

This doesn't run yet. Some more setup is required. It does a port of the
environment variables from the .travis.yml file.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>