ci: use distro's python3 package; remove pyenv python
The minimum version that gets installed with the current setup is
Python 3.5, and the only distro we have now that has such an old version is
Ubuntu 16.04.

Ubuntu 16.04 is getting a little old, so we can probably start skipping
docs install for this distro.

CentOS 7 has Python 3.6

All other distros should have something at least as new.
Removing this reduces some clutter in the build scripts.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>