ci,azure-pipelines: add windows builds
This change adds a minimal configuration to build some Windows artifacts
for libiio.
These are stored after the pipeline is done.

Adapted from Travis Collins' work.
Added into PowerShell scripts. Ideally we should not add any script-code
into yaml files, as that makes it harder to scale, and even harder to
migrate (if ever will be needed again).

Publishing artifacts, the generated DLL, static library, and the iio.h
header, which will be used in the next jobs.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>