python: Solved Python bindings Context destructor bug.
The Device class was only holding a weakref of the Context object. This could lead to calling the Context's destructor before removing all its references.
An example of crash can be found in the following issue: https://github.com/analogdevicesinc/libiio/issues/628.
Weakref was permitting Python to call the Context destructor even if it was referenced in the device object. Changing the weakref to a full reference should solve the issue.

Signed-off-by: Cristi Iacob <cristian.iacob@analog.com>