ci,linux: don't install sphinx docs on ARM{64}
Since the environment for building on ARM/ARM64 is emulated, execution is a
bit slow.
The whole build with docs takes 13 minutes on ARM and 17 minutes on ARM64.
So, for now we don't build docs on these platforms.

The way this is done, is to use a CI_BUILD_SPHINX_DOCS env var that is set
in a configuration/yaml file. This gives a bit more control over the
generation of these docs, because it signals that we want these docs to
build, and fail if we can't install dependencies for these (vs just
checking for Python versions and silently failing if it's not ok).

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>