serial: Detect timeout when reading a byte from serial port
The descrpition of sp_blocking_read_next() specifies that 0 is returned on timeout. This commit handles this.
An issue has been observed when trying to create a context for the wrong serial port (where no tiny-iiod exists on the other end) which causes a really long wait time because sp_blocking_read_next() was being called for at least 1024 times and each time it timedout after 1 second.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>