iio: wrap iio device attributes in an iio_dev_attr struct
There's quite a bit of duplication that is done with these 3 attribute
types. If we wrap them under a struct, this can be reduced.

This is useful if we want to add more buffers per IIO device. As we will
need to keep multiples of IIO buffer attributes (for each buffer).

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>