local: introduce support for new iio buffer interfaces
This change introduces support for the new IIO multibuffer support (i.e
multiple IIO buffers per one IIO device). The way this works is that
instead of using the FD for /dev/iio:deviceX to read data from it, we
request a new FD from which we read data.

This means that we can close the old FD (for /dev/iio:deviceX) and use the
new FD (from the ioctl()).

In the kernel, the new ioctls for the high-speed mmap interface should also
be at around the same time with this new multibuffer support. The ioctls
for the mmap interface should be done on the FD returned by the GETFD
ioctl.

For now, this interface is still used to access just a single IIO buffer
per device. This patch just does the transition to the new interface.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>