utilities: Implement real function iio_snprintf
Everything in libiio expects iio_snprintf() to return an error when the
data to be written overflows the buffer, but it is not how snprintf()
behaves; instead, it returns the number of characters that would have
been written if there was enough space in the buffer.

Therefore, to know that the data was truncated, it is required to
compare the return value with the size of the buffer, and nothing in
libiio does that.

Address this issue by returning -ERANGE if an overflow is detected.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>