IIOD: Add local version of iio_device_get_sample_size_mask()
The iio_device_get_sample_size_mask() is a function exported by the
libiio library, but not present in the public header. It was added only
to be used by IIOD.

Add a local version of iio_device_get_sample_size_mask() in IIOD which
only depends on the official public API, so that the libiio function can
be retired.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>