IIOD: Don't resize buffer before writing samples
IIOD supported using one single IIO buffer for both receiving and
sending samples.

While this worked *in theory*, it has never really been tested in that
configuration, as the associated IIO devices seem to always be either
receive-only, or transmit-only.

As it does rely on libiio internals, with no way to re-implement cleanly
on top of the public libiio API, disable this functionality for now.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>