Define tokens iio_api_{import,export} and __iio_api in <iio.h>
__api is defined at the top of <iio.h> and undefined at the bottom.

This makes it impossible for the other headers to use the token without
redefining __api.

Now, <iio.h> will define __iio_api according to the wanted visibility,
and define __api to __iio_api. Like before, __api will be undefined at
the end of <iio.h>, but __iio_api won't, and can be reused in other
headers.

The __iio_api_import and __iio_api_export tokens can be used to force
the importation/exportation of a symbol independently of the value of
the LIBIIO_EXPORTS macro.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>