CMake: Add support for dynamically-loaded module backends
If enabled with the WITH_MODULES CMake option, this new functionality
will support loading IIO backends built as modules from the disk.

When requesting a backend with e.g. the URI "mybackend:test", libiio
will first try to match "mybackend:" with any of its built-in backend,
then try to load /usr/lib/libiio/libiio-mybackend.so. If found, it will
search for the symbol 'iio_mybackend_backend', and if present, will
assume it is a 'struct iio_backend' and call its .create() callback.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>