CMake: Support building the serial backend as module
The serial backend can be built completely relying on the high-level
libiio and iio-backend/iiod-client APIs.

This commit introduces the new CMake option WITH_SERIAL_BACKEND_DYNAMIC.
If disabled, the serial backend will be buil-in. If enabled, the serial
backend will be built as an installable shared object, that can then be
dynamically loaded by the library when the "serial:" URI is used.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>