CMake: Don't compile with -D_GNU_SOURCE
We do not want to compile with GNU-specific extensions, as we want the
code to be as portable as possible.

Thanksfully we don't actually need _GNU_SOURCE to be defined, removing
it only triggers an error in network-unix.c about an undefined
structure, which we can fix by including the right header.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>