CI: Update build script for Linux
Enable ZSTD support, except on Ubuntu 16.04 where it doesn't build, and
explicitely set the build options as it is required now.

Note that the serial backend is disabled on Ubuntu 16.04 as its version
of libserialport is too old, and USB support in IIOD is disabled in
CentOS 7 as its kernel headers are too old.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>