CMake: Support building the USB backend as module
The USB backend can be built completely relying on the high-level
libiio and iio-backend/iiod-client APIs.

This commit introduces the new CMake option WITH_USB_BACKEND_DYNAMIC.
If disabled, the USB backend will be built-in. If enabled, the USB
backend will be built as an installable shared object, that can then be
dynamically loaded by the library when the "usb:" URI is used.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>