CMake: Support building the network backend as module
The network backend can be built completely relying on the high-level
libiio and iio-backend/iiod-client APIs.

This commit introduces the new CMake option
WITH_NETWORK_BACKEND_DYNAMIC. If disabled, the network backend will be
built-in. If enabled, the network backend will be built as an
installable shared object, that can then be dynamically loaded by the
library when the "ip:" URI is used.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>