network: Remove .get_buffer
Remove the experimental .get_buffer() callback on the network backend.

This has always been disabled by default on the config, and didn't
really prove to be that much better than the regular interface.

While it does work, it's a lot of complexity for a redundant
functionality, so we're better off without it.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>