Add IIOD responder
Add a new iiod_responder class. This new object implements the two-way
communication between an initiator (sending commands) and a recipient
(sending responses).

This will then make it possible to implement an asynchronous binary
protocol between the IIOD server and the clients. The IIOD responder is
implemented as a library, that does not depend on the libiio internals,
and can therefore be used by both sides (IIOD and clients) to handle the
communication.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>