iiod-client: Set iiod_client_pdata pointer on creation
Now that all backends use multiple instances of iiod-client, we don't
need each function to receive a pointer to the iiod_client_pdata, since
we know that for a given iiod-client instance, the iiod_client_pdata
passed as argument will always be the same.

Therefore, the iiod_client_pdata pointer can be passed as argument once
when creating the iiod-client context, and it can be kept within the
iiod_client structure, so that it doesn't have to be passed around
everytime.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>