usb: when scanning provide option to qualify USB vendor ID or product ID
When we do a USB scan, we open each device looking for a string "IIO",
and if it is not, then we close the device. The archirecture of libusb
is that when we are doing this, other applications are locked out from
using that device. This is known to cause some user problems on other
devices (not IIO) - most notibly HackRF and USRP.

This change adds the ability to restricting scans to certain vendors, or
specific product IDs - which is what most purpose build software is
looking for. (GNU Radio doesn't want to open every device, only the ones
it knows about).

This should allow us to play nicer with others. Examples:
 usb,local
 local,usb=0456:b673,usb=0456:b672
 usb=0456:*

VENDOR ID and PRODUCT ID are hexadecimal numbers (no prefix needed),
"*" (match any). By default (no vid:pid provided) all devices are opened
and checked (which is the previous behavour).

Signed-off-by: Robin Getz <robin.getz@analog.com>
Signed-off-by: Paul Cercueil <paul@crapouillou.net>